GitHub Sync (#96)
* Allow time series chart to use bar renderer

PiperOrigin-RevId: 206843411

* Fix casting error in linear scale viewport.

PiperOrigin-RevId: 207006324

* Adds a knob for flipping the vertical axis so that larger numbers go down instead of the default. KI: bar_renderer apparently doesn't support negative bars

PiperOrigin-RevId: 207162102

* BarTarggetLineRendererConfig modified to accept type other than string,
so that bar renderer could be used in a non ordinal axis.

PiperOrigin-RevId: 207166249

* Pass in the scale step size for static tick provider.

PiperOrigin-RevId: 207172296

* Make initial hint behavior assert less restrictive: allow axes that
have a rangeband config other than none.

PiperOrigin-RevId: 207180493

* Allow Dart2

PiperOrigin-RevId: 207589200

+++ charts_common/lib/src/chart/cartesian/axis/linear/linear_scale_function.dart
@@ -156,7 +156,7 @@ class LinearScaleFunction {
       switch (stepSizeConfig.type) {
         case StepSizeType.autoDetect:
           double minimumDetectedDomainStep =
-              domainInfo.minimumDetectedDomainStep;
+              domainInfo.minimumDetectedDomainStep.toDouble();
           if (minimumDetectedDomainStep != null &&
               minimumDetectedDomainStep.isFinite) {
             scalingFactor = viewportSettings.scalingFactor *
